+++ R/class_active.R
@@ -201,7 +201,6 @@ active_class <- R6::R6Class(
       } else if (target_should_run(target, meta)) {
         self$skipping <- inherits(target, "tar_pattern")
         self$flush_upload_meta_file(target)
-        target_gc(target)
         .subset2(self, "run_target")(target)
       } else {
         .subset2(self, "skip_target")(target)