+++ tests/testthat/test-tar_option_set.R
@@ -165,11 +165,11 @@ tar_test("memory", {
 })
 
 tar_test("garbage_collection", {
-  expect_equal(tar_option_get("garbage_collection"), 1000L)
+  expect_equal(tar_option_get("garbage_collection"), 0L)
   tar_option_set(garbage_collection = 5L)
   expect_equal(tar_option_get("garbage_collection"), 5L)
   tar_option_reset()
-  expect_equal(tar_option_get("garbage_collection"), 1000L)
+  expect_equal(tar_option_get("garbage_collection"), 0L)
   expect_error(
     suppressWarnings(tar_option_set(garbage_collection = "abc")),
     class = "tar_condition_validate"