+++ R/tar_outdated.R
@@ -41,7 +41,8 @@
 #'       with informative detail. Uses a `cli` progress bar.
 #'       To the right of the progress bar is a text string like
 #'       "22.6s, 4510+, 124-" (22.6 seconds elapsed, 4510 targets
-#'       detected as outdated so far, 124 targets detected as up to date so far).
+#'       detected as outdated so far,
+#'       124 targets detected as up to date so far).
 #'     * `"silent"`: print nothing.
 #' @inheritParams tar_validate
 #' @examples